Store worker compliance checks, separate models into categories
