+++ inference/server/oasst_inference_server/models/__init__.py
@@ -1,0 +1,15 @@
+from .chat import DbChat, DbMessage, DbReport, DbVote
+from .user import DbUser
+from .worker import DbWorker, DbWorkerComplianceCheck, DbWorkerEvent, WorkerEventType
+
+__all__ = [
+    "DbChat",
+    "DbMessage",
+    "DbReport",
+    "DbVote",
+    "DbUser",
+    "DbWorker",
+    "DbWorkerComplianceCheck",
+    "DbWorkerEvent",
+    "WorkerEventType",
+]
