+++ inference/server/oasst_inference_server/models/chat.py
@@ -1,12 +1,11 @@
 import datetime
-import enum
 from uuid import uuid4
 
 import sqlalchemy as sa
 import sqlalchemy.dialects.postgresql as pg
 from oasst_inference_server.schemas import chat as chat_schema
 from oasst_shared.schemas import inference
-from sqlmodel import Field, Index, Relationship, SQLModel
+from sqlmodel import Field, Relationship, SQLModel
 
 
 class DbMessage(SQLModel, table=True):
@@ -86,42 +85,3 @@ class DbReport(SQLModel, table=True):
 
     def to_read(self) -> inference.Report:
         return inference.Report(id=self.id, report_type=self.report_type, reason=self.reason)
-
-
-class WorkerEventType(str, enum.Enum):
-    connect = "connect"
-
-
-class DbWorkerEvent(SQLModel, table=True):
-    __tablename__ = "worker_event"
-
-    id: str = Field(default_factory=lambda: str(uuid4()), primary_key=True)
-    worker_id: str = Field(foreign_key="worker.id", index=True)
-    worker: "DbWorker" = Relationship(back_populates="events")
-    time: datetime.datetime = Field(default_factory=datetime.datetime.utcnow)
-    event_type: WorkerEventType
-    worker_config: inference.WorkerConfig | None = Field(None, sa_column=sa.Column(pg.JSONB))
-
-
-class DbWorker(SQLModel, table=True):
-    __tablename__ = "worker"
-
-    id: str = Field(default_factory=lambda: str(uuid4()), primary_key=True)
-    api_key: str = Field(default_factory=lambda: str(uuid4()), index=True)
-    name: str
-
-    in_compliance_check: bool = Field(default=False, sa_column=sa.Column(sa.Boolean, server_default=sa.text("false")))
-    next_compliance_check: datetime.datetime | None = Field(None)
-    events: list[DbWorkerEvent] = Relationship(back_populates="worker")
-
-
-class DbUser(SQLModel, table=True):
-    __tablename__ = "user"
-    __table_args__ = (Index("provider", "provider_account_id", unique=True),)
-
-    id: str = Field(default_factory=lambda: str(uuid4()), primary_key=True)
-
-    provider: str = Field(index=True)
-    provider_account_id: str = Field(index=True)
-
-    display_name: str = Field(nullable=False, max_length=256)
